WIP program with objects.
